requirements.txt
pyttsx3
pywin32
pyaudio
SpeechRecognition
google-generativeai